+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Change Log
 
 HEAD
 ----
+-   Add a new `max_connections` option for the `Configuration` class, to increase the session poolsize
+    on a per-server, per-credentials basis. Useful when exchangelib is used with threads, where one may
+    wish to increase the number of concurrent connections to the server.
 
 
 3.3.2